+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `Stickers` test cases
 - `StickerOwnerUserId` parameter to Systems Integration Tests configs
 - Exception `InvalidStickerSetNameException`
+- Exception `InvalidStickerEmojisException`
 
 ## [13.4.0-rc2] - 2017-11-26